extend magiskhide with exec option
this shall be used to start a process in "stock" environment
where magisk components are not visible

this is needed in very early boot phase to start special root
detection processes like sony xperia ric daemon

when ric daemon thinks bootloader is still locked and detects
presence of /sbin/su or other utilities, kernel is rebooted
in it's very early boot phase

this patch adds new --exec option which will create private
mount namespace and change all mount points sharing to MS_PRIVATE
so hide_daemon() can be used to unmount magisk mounts just before
exec() to target process

+++ native/jni/magiskhide/proc_monitor.cpp
@@ -108,17 +108,9 @@ static bool is_pid_safetynet_process(const int pid) {
 	return !strcmp(buf, SAFETYNET_PROCESS);
 }
 
-static void hide_daemon(int pid) {
-	LOGD("hide_daemon: handling pid=[%d]\n", pid);
-
+void do_hide_daemon(int pid) {
 	char buffer[4096];
 	vector<string> mounts;
-
-	manage_selinux();
-	clean_magisk_props();
-
-	if (switch_mnt_ns(pid))
-		goto exit;
 
 	snprintf(buffer, sizeof(buffer), "/proc/%d", pid);
 	chdir(buffer);
@@ -144,8 +136,20 @@ static void hide_daemon(int pid) {
 			lazy_unmount(buffer);
 		}
 	}
+}
 
-exit:
+
+static void hide_daemon(int pid) {
+	LOGD("hide_daemon: handling pid=[%d]\n", pid);
+
+	manage_selinux();
+	clean_magisk_props();
+
+	if (switch_mnt_ns(pid))
+		return;
+
+	do_hide_daemon(pid);
+
 	// Send resume signal
 	kill(pid, SIGCONT);
 	_exit(0);
